Refactor ShoppingCart Scene and Cell to use true viewModel
