+++ FakeNFT/Foundation/Coordinator.swift
@@ -1,0 +1,3 @@
+protocol Coordinator {
+    func start()
+}
