--- FakeNFT/Scenes/ShoppingCart/Cart/Dependencies.swift
@@ -1,6 +1,0 @@
-extension ShoppingCartController {
-    struct Dependencies {
-        let nftProvider: NftProvider
-        let shoppingCart: ShoppingCart
-    }
-}
