+++ FakeNFT/Scenes/ShoppingCart/ShoppingCart/NftDelete/NftDeleteView.swift
